Convert a bunch of printf from %d to %ld to match the NumPart extension.

+++ libgadget/gravpm.c
@@ -147,7 +147,7 @@ static PetaPMRegion * _prepare(PetaPM * pm, PetaPMParticleStruct * pstruct, void
     MPI_Reduce(&r, &maxNregions, 1, MPI_INT, MPI_MAX, 0, MPI_COMM_WORLD);
     message(0, "max number of regions is %d\n", maxNregions);
 
-    int i;
+    int64_t i;
     int numswallowed = 0;
     #pragma omp parallel for reduction(+: numswallowed)
     for(i =0; i < PartManager->NumPart; i ++) {
@@ -162,7 +162,7 @@ static PetaPMRegion * _prepare(PetaPM * pm, PetaPMParticleStruct * pstruct, void
     }
 
     /* now lets mark particles to their hosting region */
-    int numpart = 0;
+    int64_t numpart = 0;
 #pragma omp parallel for reduction(+: numpart)
     for(r = 0; r < *Nregions; r++) {
         regions[r].numpart = pm_mark_region_for_node(regions[r].no, r, pstruct->RegionInd, tree);
@@ -175,7 +175,7 @@ static PetaPMRegion * _prepare(PetaPM * pm, PetaPMParticleStruct * pstruct, void
     }
     /* All particles shall have been processed just once. Otherwise we die */
     if((numpart+numswallowed) != PartManager->NumPart) {
-        endrun(1, "Processed only %d particles out of %d\n", numpart, PartManager->NumPart);
+        endrun(1, "Processed only %ld particles out of %ld\n", numpart, PartManager->NumPart);
     }
     for(r =0; r < *Nregions; r++) {
         convert_node_to_region(pm, &regions[r], tree->Nodes);
